Hardware volume buttons change pages.
First instinct was that 'Up' means 'Next', but after reviewing
some other book viewer apps and experimentation, it seems to
work better the other way around. This is because if you are
reading, e.g. a PDF with many pages, then you would *scroll down*
on the screen to go to the next page, so *volume down* also does this.